# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FlowVerse 3D HTTP API
  version: 0.1.0
  summary: API para diseño, versionado, validación y simulación de flujos.
  description: |
    Contrato HTTP canónico del MVP. La autenticación utiliza cookies HttpOnly
    opacas. Toda mutación autenticada debe incluir `X-CSRF-Token`. Los errores
    comparten un sobre estable y cada respuesta incluye `X-Request-Id`.
servers:
  - url: http://localhost:8080
    description: Desarrollo local
...
security:
  - cookieAuth: []
paths:
...
  /v1/organizations:
    get:
      tags: [Organizations]
      operationId: listOrganizations
      summary: Lista las organizaciones donde el usuario tiene membresía activa.
      responses:
        "200":
          description: Organizaciones visibles.
          content:
            application/json:
              schema:
                $ref: "./schemas/enterprise.schema.json#/$defs/OrganizationList"
    post:
      tags: [Organizations]
      operationId: createOrganization
      summary: Crea una organización y asigna al usuario actual como owner activo.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/enterprise.schema.json#/$defs/OrganizationCreate"
      responses:
        "201":
          description: Organización creada y auditada atómicamente.
          content:
            application/json:
              schema:
                $ref: "./schemas/enterprise.schema.json#/$defs/Organization"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/organizations/{organizationId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [Organizations]
      operationId: getOrganization
      summary: Obtiene una organización visible para un miembro activo.
      responses:
        "200":
          description: Organización.
          content:
            application/json:
              schema:
                $ref: "./schemas/enterprise.schema.json#/$defs/Organization"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/organizations/{organizationId}/members:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [Organizations]
      operationId: listOrganizationMembers
      summary: Lista miembros; disponible para owner, admin y auditor activos.
      responses:
        "200":
          description: Miembros de la organización.
          content:
            application/json:
              schema:
                $ref: "./schemas/enterprise.schema.json#/$defs/OrganizationMemberList"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Organizations]
      operationId: setOrganizationMember
      summary: Crea o actualiza por email una membresía registrada.
      description: Un admin no puede asignar ni modificar owners; el último owner activo nunca puede ser degradado o suspendido.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/enterprise.schema.json#/$defs/OrganizationMemberSet"
      responses:
        "200":
          description: Membresía actualizada.
          content:
            application/json:
              schema:
                $ref: "./schemas/enterprise.schema.json#/$defs/OrganizationMemberView"
        "201":
          description: Membresía creada.
          content:
            application/json:
              schema:
                $ref: "./schemas/enterprise.schema.json#/$defs/OrganizationMemberView"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /v1/organizations/{organizationId}/audit:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [Enterprise]
      operationId: listOrganizationAudit
      summary: Lista la cadena append-only con paginación por secuencia.
      parameters:
        - $ref: "#/components/parameters/EnterpriseAuditAfterSequence"
        - $ref: "#/components/parameters/EnterpriseAuditLimit"
      responses:
        "200":
          description: Página estable de auditoría.
          content:
            application/json:
              schema:
                $ref: "./schemas/enterprise.schema.json#/$defs/AuditEventList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/organizations/{organizationId}/audit/verify:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [Enterprise]
      operationId: verifyOrganizationAudit
      summary: Verifica hashes, secuencias y checkpoint contra truncamiento de cola.
      responses:
        "200":
          description: Resultado de integridad; una cadena inválida también responde 200 con valid=false.
          content:
            application/json:
              schema:
                $ref: "./schemas/enterprise.schema.json#/$defs/AuditVerification"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/organizations/{organizationId}/projects:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [Organizations]
      operationId: listOrganizationProjects
      summary: Lista proyectos adscritos a la organización.
      responses:
        "200":
          description: Proyectos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/organizations/{organizationId}/projects/{projectId}/attach:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/ProjectId"
    post:
      tags: [Organizations]
      operationId: attachProjectToOrganization
      summary: Adscribe un proyecto sin permitir moverlo entre tenants.
      description: Requiere owner/admin activo de la organización y rol owner sobre el proyecto.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Proyecto adscrito.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flowverse_access
...
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    EnterpriseAuditAfterSequence:
      name: afterSequence
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    EnterpriseAuditLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 100
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
...
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
    Issue:
      type: object
      additionalProperties: false
      required: [code, severity, message]
      properties:
        code:
          type: string
        severity:
          enum: [info, warning, error]
        message:
          type: string
        path:
          type: string
        nodeId:
          type: string
        edgeId:
          type: string
...
    Project:
      type: object
      additionalProperties: false
      required: [id, name, role, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 2000
        role:
          enum: [owner, editor, viewer]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Project"
...
  responses:
...
    BadRequest:
      description: |
        Solicitud inválida. En creación de runs incluye
        `idempotency.key_required`, `idempotency.key_invalid`, JSON inválido u
        overrides inexistentes.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Recurso inexistente o no visible para el usuario.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Estado actual incompatible con la operación.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationFailed:
      description: Entrada o grafo inválido.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: Límite temporal excedido.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ServiceUnavailable:
      description: Dependencia requerida no disponible.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"